make it work for iphone 6s and iphone 6s plus
While testing it, I have encountered that it didn't work for iPhone 6S and 6S Plus, due to weak taptic engine that doesn't support all possible variants that work for >= iPhone 7. So I have added haptic feedback for some of the variants such as `selection`, `impactMedium` and `notificationWarning`.

This article have lead me to this -> https://medium.com/@sdrzn/make-your-ios-app-feel-better-a-comprehensive-guide-over-taptic-engine-and-haptic-feedback-724dec425f10